+++ FlexSprinkle/Models/CoreData/WorkoutSet+CoreDataProperties.swift
@@ -21,7 +21,7 @@ extension WorkoutSet {
     @NSManaged public var reps: Int32
     @NSManaged public var setNumber: Int32
     @NSManaged public var time: Int32
-    @NSManaged public var weight: Int32
+    @NSManaged public var weight: Float
     @NSManaged public var isCompleted: Bool
     @NSManaged public var setIndex: Int32
     @NSManaged public var details: WorkoutDetail?
