+++ FlexSprinkle/Models/WorkoutDetailsInput.swift
@@ -19,14 +19,14 @@ struct WorkoutDetailInput {
 struct SetInput: Identifiable {
     var id: UUID? // Make sure this property exists to match your function's expectations
     var reps: Int32
-    var weight: Int32
+    var weight: Float
     var time: Int32
     var distance: Float
     var isCompleted: Bool
     var setIndex: Int32
     
     // Initialize with default values
-    init(id: UUID? = nil, reps: Int32 = 0, weight: Int32 = 0, time: Int32 = 0, distance: Float = 0, isCompleted: Bool = false, setIndex: Int32 = 0)
+    init(id: UUID? = nil, reps: Int32 = 0, weight: Float = 0, time: Int32 = 0, distance: Float = 0, isCompleted: Bool = false, setIndex: Int32 = 0)
     {
         self.id = id
         self.reps = reps
